refactor: move the Header's components to a separate folder
