+++ src/components/HeadersComponents/Signbar/Signbar.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import CustomList from '../CustomList/CustomList';
+import CustomList from '../../CustomList/CustomList';
 import SignbarItems from '../SignbarItems/SignbarItems';
-import CLASSNAMES from '../../utils/scripts/classnames';
+import CLASSNAMES from '../../../utils/scripts/classnames';
 
 function Signbar({ сlassNameFromParent }) {
   const {
