+++ src/pages/Main.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Header from '../components/Header/Header';
+import Header from '../components/HeadersComponents/Header/Header';
 import Content from '../components/Content/Content';
 import Promo from '../components/Promo/Promo';
 import History from '../components/History/History';
